FIX: store Discourse topic ids as bigint

+++ db/migrate/20260831214000_create_community_topic_votes.rb
@@ -4,7 +4,7 @@ class CreateCommunityTopicVotes < ActiveRecord::Migration[7.2]
   def change
     create_table :discourse_community_platform_votes do |t|
       t.integer :community_id, null: false
-      t.integer :topic_id, null: false
+      t.bigint :topic_id, null: false
       t.integer :user_id, null: false
       t.integer :value, null: false, limit: 2
       t.timestamps null: false
@@ -27,7 +27,7 @@ class CreateCommunityTopicVotes < ActiveRecord::Migration[7.2]
 
     create_table :discourse_community_platform_topic_scores do |t|
       t.integer :community_id, null: false
-      t.integer :topic_id, null: false
+      t.bigint :topic_id, null: false
       t.integer :upvotes, null: false, default: 0
       t.integer :downvotes, null: false, default: 0
       t.integer :score, null: false, default: 0
